Move back to 3.0 (native) source format.
When trying to find a good git-buildpackage repository layout for the non-native
package, I realized that all this is highly Debian-specific, and one often needs
to make changes on the "upstream" and "Debian" bits at the same time. When this
happens, one can't be bothered to have to put out a new "upstream release"
first, and then importing it into the packaging branch, before being able to do
the needed changes.

Accordingly remove .gitattributes and debian/watch, that were only needed for
a non-native package.